requirements.txt
pandas
streamlit
google-generativeai
python-dotenv